Moved PcaAddressFieldMapping constants to main module.
