+++ src/PcaAddressFieldMapping/PcaAddressMode.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\pca_address\FieldMapping;
+namespace Drupal\loqate\PcaAddressFieldMapping;
 
 /**
  * Field mapping mode for PCA Address.
